Refactored Bitbucket tests for DRYness, reusability
